refactor: 어드민 - 쿠폰 관심사 리팩토링
