+++ src/refactoring/components/CouponAdmin.tsx
@@ -1,0 +1,86 @@
+import { useState } from "react";
+import { Coupon } from "../../types";
+import { useNewCoupon } from "../hooks/useNewCoupon";
+import { CouponCard } from "./CouponCard";
+
+interface Props {
+  coupons: Coupon[];
+  onCouponAdd: (newCoupon: Coupon) => void;
+}
+
+export const CouponAdmin = ({ coupons, onCouponAdd }: Props) => {
+  const {
+    newCoupon,
+    resetNewCoupon,
+    setNewCouponCode,
+    setNewCouponDiscountType,
+    setNewCouponDiscountValue,
+    setNewCouponName,
+  } = useNewCoupon();
+
+  const handleAddCoupon = () => {
+    onCouponAdd(newCoupon);
+    resetNewCoupon();
+  };
+
+  return (
+    <div>
+      <h2 className="text-2xl font-semibold mb-4">쿠폰 관리</h2>
+      <div className="bg-white p-4 rounded shadow">
+        <div className="space-y-2 mb-4">
+          <input
+            type="text"
+            placeholder="쿠폰 이름"
+            value={newCoupon.name}
+            onChange={(e) => setNewCouponName(e.target.value)}
+            className="w-full p-2 border rounded"
+          />
+          <input
+            type="text"
+            placeholder="쿠폰 코드"
+            value={newCoupon.code}
+            onChange={(e) => setNewCouponCode(e.target.value)}
+            className="w-full p-2 border rounded"
+          />
+          <div className="flex gap-2">
+            <select
+              value={newCoupon.discountType}
+              onChange={(e) =>
+                setNewCouponDiscountType(
+                  e.target.value as "amount" | "percentage"
+                )
+              }
+              className="w-full p-2 border rounded"
+            >
+              <option value="amount">금액(원)</option>
+              <option value="percentage">할인율(%)</option>
+            </select>
+            <input
+              type="number"
+              placeholder="할인 값"
+              value={newCoupon.discountValue}
+              onChange={(e) =>
+                setNewCouponDiscountValue(parseInt(e.target.value))
+              }
+              className="w-full p-2 border rounded"
+            />
+          </div>
+          <button
+            onClick={handleAddCoupon}
+            className="w-full bg-green-500 text-white p-2 rounded hover:bg-green-600"
+          >
+            쿠폰 추가
+          </button>
+        </div>
+        <div>
+          <h3 className="text-lg font-semibold mb-2">현재 쿠폰 목록</h3>
+          <div className="space-y-2">
+            {coupons.map((coupon, index) => (
+              <CouponCard coupon={coupon} index={index} key={index} />
+            ))}
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
